requirements.txt
asgiref==3.7.2
black==24.1.1
click==8.1.7
Django==5.0.1
django-redis==5.4.0
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
python-dotenv==1.0.1
pytz==2023.4
redis==5.0.1
sqlparse==0.4.4
